+++ app/models/project.rb
@@ -6,7 +6,7 @@
 class Project < ApplicationRecord
   PER_PAGE = 40
   AUTO_LOCK_SHEETS = [
-    ["Never Lock Sheets", "never"],
+    ["Never lock sheets", "never"],
     ["After 24 hours", "after24hours"],
     ["After 1 week", "after1week"],
     ["After 1 month", "after1month"]