+++ test/system/aer/admins_test.rb
@@ -9,13 +9,6 @@ class AdminsTest < ApplicationSystemTestCase
     @review_admin = users(:aes_review_admin)
   end
 
-  test "visit adverse event dashboard as admin" do
-    visit_login(@review_admin)
-    visit ae_module_dashboard_url(@project)
-    assert_selector "h1", text: "AE Module Dashboard"
-    screenshot("visit-adverse-event-dashboard")
-  end
-
   test "visit adverse event setup designs as admin" do
     visit_login(@review_admin)
     visit ae_module_admins_setup_designs_url(@project)