+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 69.0.0
+## 69.0.0 (January 28, 2019)
 
 ### Enhancements
 - **Adverse Event Reviews Added**